[feat] Moved business logic out of FSUtil and into the NinoxProjectService
